Relocate stocktake line tests to graphql crate
